backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
